fix old test
